Add chunked request decoding to the CGI plugin
When a request using 'Transfer-Encoding: chunked' is routed to the CGI
backend, the request body is now read and sent to the script chunk by
chunk, allowing for requests with dynamic body length to pass. The CGI
process's stdin is always closed after the last chunk has been written,
regardless of --cgi-close-stdin-on-eof.

A caveat is that it's only started to read the CGI process's output
after the full request body has been consumed, meaning a CGI script may
not start writing a response before it has read the entire body to avoid
getting blocked during the write operation to stdout.

Closes #2140